+++ changelog.md
@@ -1,5 +1,21 @@
 # Change Log
 
+# v1.4 -> v1.5 Changes
+ - Calculator Rework
+     - Now uses floats because the Batch Exporter will soon allow non-whole number scale percentages.
+     - Modifier key increment system: Hold a key to change the scale box's arrow increments.
+         - Alt: 0.01%
+         - Shift: 0.1%
+         - Default: 1.0%
+         - Ctrl: 10%
+     - Width and Height numbers are now in editable text boxes.
+         - Edit Width: Scale and Height will automatically update.
+         - Edit Height: Scale and Width will automatically update.
+     - Scale limit increased from 100% to 200% to allow dimension calculation for small upscaling.
+     - Added a tooltip.
+ - Added [xalign x yalign y] as an export option.
+ - Layer name parsing now uses a dictionary system to be more \
+robust and better for additional tag implementations.
 # v1.3 -> v1.4 Changes
  - `Rename Batch-Exported Files` is now a button instead of a checkbox.
  - `Rename Batch-Exported Files` now works with layers that were exported as groups into folders.